printer, editable payslip and icons

+++ src/components/EmployeeTable.jsx
@@ -1,16 +1,21 @@
 'use client'
-import React, { useRef, useState } from 'react'
+import React, { useState } from 'react'
 import { Table } from 'flowbite-react';
 import useSWR, { mutate } from 'swr';
 import Link from 'next/link';
 import { useSession } from 'next-auth/react';
 import Image from 'next/image';
+import Popup from './Popup';
 
 const EmployeeTable = () => {
 
   const session = useSession();
   const apiUrl = process.env.NEXT_PUBLIC_SITE_URL;
   const companyName = session.data.user.name.companyName;
+
+    const [open, setOpen] = useState(false);
+    const [employeeId, setEmployeeId] = useState('');
+
     // fetch employees
   const fetcher = (...args) => fetch(...args).then((res) => res.json());
 
@@ -24,21 +29,22 @@ const EmployeeTable = () => {
     return <div>Error: {error.message}</div>;
   }
 
-  const handleDelete = async(employeeId) => {
-    try {
-        const res = await fetch(`${apiUrl}/api/employees/${employeeId}`, {
-            method: 'DELETE',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-        });
-        console.log(res.status);
-        res.status === 200;
-        mutate('/api/employees');
-    } catch(error) {
-        console.log(error);
+  const handleDelete = async (canDelete) => {
+    if (canDelete) {
+        try {
+            const res = await fetch(`${apiUrl}/api/employees/${employeeId}`, {
+                method: 'DELETE',
+                headers: {
+                  'Content-Type': 'application/json',
+                },
+              });
+              res.status === 200;
+              mutate(`${apiUrl}/api/employees/companyName/${companyName}`);
+        } catch (error) {
+          console.log(error);
+        }
     }
-  }
+  };
 
   
 
@@ -97,7 +103,7 @@ const EmployeeTable = () => {
                     <Table.Cell className="whitespace-nowrap font-medium text-gray-900 dark:text-white">{x.NRIC}</Table.Cell>
                     <Table.Cell className="whitespace-nowrap font-medium text-gray-900 dark:text-white">{new Date(x.dateOfBirth).toLocaleDateString('en-GB')}</Table.Cell>
                     <Table.Cell className="whitespace-nowrap font-medium text-gray-900 dark:text-white">{x.designation}</Table.Cell>
-                    <Table.Cell className="whitespace-nowrap font-medium text-gray-900 dark:text-white">{new Date(x.joinDate).toLocaleDateString()}</Table.Cell>
+                    <Table.Cell className="whitespace-nowrap font-medium text-gray-900 dark:text-white">{new Date(x.joinDate).toLocaleDateString('en-GB')}</Table.Cell>
                     <Table.Cell className="whitespace-nowrap font-medium text-gray-900 dark:text-white">{x.isResigned ? "Resigned" : "Active"}</Table.Cell>
                     <Table.Cell className="whitespace-nowrap font-medium text-gray-900 dark:text-white">
                         <Link href={`/dashboard/employee-list/${x.id}`} >
@@ -105,7 +111,10 @@ const EmployeeTable = () => {
                         </Link>
                     </Table.Cell>
                     <Table.Cell className="whitespace-nowrap font-medium text-gray-900 dark:text-white">
-                        <div onClick={() => handleDelete(x.id)} className="cursor-pointer">
+                        <div onClick={() => {
+                            setOpen(true)
+                            setEmployeeId(x.id)
+                        }} className="cursor-pointer">
                             <Image src="/trash.svg" alt="Edit" height={16} width={16} />
                         </div>
                     </Table.Cell>
@@ -113,7 +122,12 @@ const EmployeeTable = () => {
             ))}
 
             </Table.Body>
-        </Table>
+        </Table>     
+        {
+            open && 
+            <Popup label={"Are you sure you want to delete this employee ?"} setOpen={setOpen} fn={handleDelete}/>
+        }
+
     </div>
   )
 }
